Use StrategyHelper for safe executions and drop redundant methods
