+++ test/Polly.Core.Tests/ResilienceStrategyTests.cs
@@ -6,30 +6,6 @@ public partial class ResilienceStrategyTests
 {
     public static readonly CancellationToken CancellationToken = new CancellationTokenSource().Token;
 
-    [InlineData(true)]
-    [InlineData(false)]
-    [Theory]
-    public async Task ExecuteCallbackSafeAsync_CallbackThrows_EnsureExceptionWrapped(bool isAsync)
-    {
-        await TestUtilities.AssertWithTimeoutAsync(async () =>
-        {
-            var outcome = await ResilienceStrategy.ExecuteCallbackSafeAsync<string, string>(
-            async (_, _) =>
-            {
-                if (isAsync)
-                {
-                    await Task.Delay(15);
-                }
-
-                throw new InvalidOperationException();
-            },
-            ResilienceContextPool.Shared.Get(),
-            "dummy");
-
-            outcome.Exception.Should().BeOfType<InvalidOperationException>();
-        });
-    }
-
     [Fact]
     public void DebuggerProxy_Ok()
     {
